Increased ball swipe padding :wrench: decreased decay deceleration :sparkles:
